src/requirements/dev.txt
-r base.txt 


flake8==4.0.1
black==23.1.0
isort==5.12.0
watchfiles==0.18.1

pytest-django==4.5.2
pytest-factoryboy==2.5.1
pytest-cov==4.0.0
Faker==18.3.1
